Initial code to load an mpo file and display with 3D Vision Direct
This loads both images from an .mpo file using Pillow, sets nvapi to 3D
Vision Direct and uses the comtypes version of DirectPython to display
it.

It also has experimental code to set the necessary bits in the driver
profile to enable 3D in windowed mode and disable the stereo memo,
however this is not yet complete as it targets the base profile rather
than python, and there are a few other issues with modifying the profile
(such as requiring admin priveleges), making it a lower priority.