Add support for parallax adjustment
Holding the middle mouse button and dragging up/down will now adjust the
parallax between the left and right images. Currently the parallax
stored in the image file is ignored - it is found in the EXIF tag of the
*second* image in the .mpo file, which Pillow currently does not expose.

Alternatively, holding P and dragging left/right with either the left or
right mouse buttons will also adjust the parallax.

It also does not yet apply the same cropping as other image viewers do -
since the goal is to give control of that to the user this is a low
priority.