Add support for horizontal, vertical and depth crops
To crop an image, hold C and drag an edge with the left button to crop
horizontally or vertically, or use the right mouse button on the left or
right edges to change the depth of the crop instead.

This is approximately one million times easier than any other tool I am
aware of.