Respect current image position to select border, limit crops
If the image is panned or cropped, use it's current bounds to select
which border to adjust when cropping so that the selection feels natural
no matter what. In the event the image has been cropped to 0px in either
dimension use a fallback path that avoids a divide by zero.

Limit the crops so they cannot overlap, i.e. so that the left border
cannot be cropped further to the right than the right border.