Enable stereo in windowed mode via 3D-Hub Player profile
This is the same trick we use in many 3D Vision mods to enable stereo in
windowed mode since it can be done by the application without requiring
admin privileges. It is not an ideal solution, but it usually works well
enough, though can be broken by driver updates that add a non-stereo
profile for the application's executable name.