Add save routine
This routine saves a copy of the image as a side-by-side .jps file,
applying the cropping and parallax adjustments to each side of the image
as required. The resulting image may have black bars on either side to
achieve the correct cropping and parallax adjustments while maintaining
the constraint that both left and right images must be equal width.