Explicitly round floating point numbers when saving file
This will hopefully fix an edge case where the right image might
slightly leak into the left image because it should ensure that the
width is always rounded up, and other rounding should go to the nearest
integer.

There might still be some other off by one or rounding issues to look
out for, but hopefully will only be minor.