Make fit to window take crop into consideration
Fit to window will leave black bars for the crop, but if the user has
cropped the image then hit the key to refit to window, they probably
want it fit after their crop.

This calculates the new scale value based on the previous rectangle,
which allows us to skip the parallax and vertical offset calculations as
they are already in said rectangle.

The pan calculation is based directly off the crop values, but since the
pan is 0.0 for centered and since parallax and vertical offset leave the
image centered we can ignore them there as well.