Slow down parallax and depth crop adjustments
These adjustments involve moving the mouse perpendicular to the 2D
direction of movement on either image (because it is moving both in
opposite directions and therefore does not have a single correct
horizontal direction, so vertical makes more sense), but they can adjust
the depth very rapidly. Slow them down by a factor of 2 to make it a bit
easier to perform precise adjustments.