Add keys to open previous/next file in directory
Page Up/Down can now be used to go to the previous or next file in the
same directory. It uses some smarts to choose the "best" file if
multiple exist with the same prefix - that is, it will try to open a
previously cropped file with the highest index, and skip over lower
priority files.