Add Shift+Wheel Up/Down as alternate way to load previous/next file
This makes things faster to go between more images in the same
directory, and can be used as a decent viewer as well. The reason to
choose Shift is that we don't want the wheel by itself to load a new
image since this is a primarily an editor, not a viewer, and Ctrl+wheel
is usually used for zooming in many image viewers, and we do not have
have a good reason to change the meaning here (quite the opposite).