Save a .spct file with metadata along with the modified .jps
The .spct file contains the current cropping settings, which can then be
opened later to load the original image with the same settings
pre-applies. This allows subsequent cropping attempts to be applied
without the degradation associated with continually saving and loading a
jpeg file, allows a photo to be uncropped without having to start again
from scratch, and allows a crop to be re-applied if the source image has
been modified externally (e.g. tweaked with rawtherapee)