Don't reset output mode, etc. when loading next file
If the user has changed the output mode, or swapped eyes, they probably
want to continue using the same output format for subsequent files, so
do not reset it when loading the next file. Also leave background alone
- if a .spct file is loaded that will set the background, but if a
subsequent .jps or .mpo file is loaded it is reasonably likely the user
will want to use the same background colour they just set.