Add support for loading mono JPEG and PNG images
This will load the image as both left and right images - this allows for
some interesting scenarios. A simple consequence is that the parallax
can still be adjusted to move a 2D image to an arbitrary depth.

But the most interesting consequence is if that image was actually a
stereo image side by side, the parallax can be adjusted far enough to
stereo fuse the correct pair (the correct direction depends on whether
the image is LR or RL), then the stereo crop can be adjusted to show
only that freshly fused image. Similarly the vertical alignment could be
adjusted to fuse a stereo image stored in Top-and-Bottom format.

This may not be quite as simple as just choosing what kind of stereo
image is being loaded (and this does not rule out adding that later),
but provides a general solution to processing images taken with a Loreo
9005 lens or similar, where there is some overlap between the images
that may be desired to show on one, both or none of the images.