Fix incorrect panning on fit to window
We actually do need to take the parallax into account when calculating
the panning to fit to window, otherwise in extreme cases (e.g. after
adjusting an image from the Loreo lens) it will still be misaligned.

Split the horizontal offset calculation into two functions so the first
part can be used as a helper to calculate the minimum and maximum
positions without trimming for the panning calculation.