Allow navigation to other files with the same extension
By default SPCT will only navigate to known stereo images, but if the
user wanted to open a mono (or SBS, etc) JPEG image than they might want
to navigate to other images of the same type, so add the extension of
any manually opened files to the list of extensions to search for while
navigating.

Also prioritise stereo .jps, .mpo and .pns images over mono versions
when navigating, though the order within each set is not defined.